Enable UBIFS boot
The first 16MiB of the flash are reserved for u-boot. The kobs-ng tool
already programms the FCB and DBBT redundant (4 times) and also the u-boot
image is programmed redundant.

The reset of the nand is a UBI partition, which implements wear-leveling
over this part of the nand. On top of that we define 3 UBIFS volumes,
which have names similar like the partitions of our SD-cards:
16MiB "bootldr", 256MiB "system" and the rest "C".

The linux kernel is the uImage file in the "bootldr" volume. The rootfs
is the "system" volume. The u-boot bootcommands are programmed so that
it loads the kernel from the correct UBIFS volume and provides the
correct kernel-commandline for mounting the rootfs.